+++ components/slider/stories/slider.stories.js
@@ -1,4 +1,5 @@
 import { disableDefaultModes } from "@spectrum-css/preview/modes";
+import { Sizes } from "@spectrum-css/preview/decorators";
 import { isFocused } from "@spectrum-css/preview/types";
 import { version } from "../package.json";
 import { SliderGroup } from "./slider.test.js";
@@ -6,6 +7,17 @@ import { Template } from "./template.js";
 
 /**
  * A slider allows users to quickly select a value within a range. They should be used when the upper and lower bounds to the range are invariable.
+ * 
+ * ## Indicating focus
+ * Focus must be bubbled up to the parent so its descendants can be styled.
+ *
+ * Thus, implementations should add the following class to the `.spectrum-Slider` parent class in the following situations:
+ *
+ * - `.is-disabled`: when the slider is disabled
+ * 
+ * Implementations should also bubble the following class to the `.spectrum-Slider-controls` parent class in the following situations:
+
+ * - `.is-focused`: when the handle input is focused with the mouse or keyboard
  */
 export default {
 	title: "Slider",
@@ -97,6 +109,17 @@ export default {
 			control: "boolean",
 			if: { arg: "variant", neq: "ramp" },
 		},
+		showTickLabels: {
+			name: "Show associated tick labels",
+			description: "Displays the values at each step of the slider",
+			type: { name: "boolean" },
+			table: {
+				type: { summary: "boolean" },
+				category: "Component",
+			},
+			control: "boolean",
+			if: { arg: "showTicks", truthy: true },
+		},
 		isDisabled: {
 			name: "Disabled",
 			type: { name: "boolean" },
@@ -117,6 +140,7 @@ export default {
 		isDisabled: false,
 		isFocused: false,
 		showTicks: false,
+		showTickLabels: false,
 		labelPosition: "top",
 		label: "Slider label",
 		size: "m",
@@ -136,10 +160,45 @@ export default {
 	},
 };
 
+/**
+ * Sliders should always have a label. In rare cases where context is sufficient and an accessibility expert has reviewed the design, the label could be undefined. Top labels are the default and are recommended because they work better with long copy, localization, and responsive layouts.
+ */
 export const Default = SliderGroup.bind({});
 Default.args = {};
 
 // ********* DOCS ONLY ********* //
+/**
+ * If a slider's label is undefined, it should still include an aria-label in HTML (depending on the context, “aria-label” or “aria-labelledby”).
+ */
+export const WithoutLabel = Template.bind({});
+WithoutLabel.args = { 
+	label: "",
+};
+WithoutLabel.tags = ["!dev"];
+WithoutLabel.storyName = "Without label";
+WithoutLabel.parameters = {
+	chromatic: {
+		disableSnapshot: true,
+	},
+};
+
+export const Sizing = (args, context) => Sizes({
+	Template,
+	withBorder: false,
+	withHeading: false,
+	...args,
+}, context);
+Sizing.args = Default.args;
+Sizing.tags = ["!dev"];
+Sizing.parameters = {
+	chromatic: {
+		disableSnapshot: true,
+	},
+};
+
+/**
+ * The track of the slider can have a fill. By default, the fill originates from the left side of the track.
+ */
 export const Filled = Template.bind({});
 Filled.args = {
 	...Default.args,
@@ -152,6 +211,9 @@ Filled.parameters = {
 	},
 };
 
+/**
+ * With fill and offset. If the value represents an offset, the fill start can be set to represent the point of origin. This allows the slider fill to start from inside the track.
+ */
 export const FilledOffset = Template.bind({});
 FilledOffset.args = {
 	...Default.args,
@@ -177,6 +239,9 @@ Ramp.parameters = {
 	},
 };
 
+/**
+ * A range slider with two handles.
+ */
 export const Range = Template.bind({});
 Range.args = {
 	...Default.args,
@@ -189,10 +254,12 @@ Range.parameters = {
 	},
 };
 
+/**
+ * Spectrum tick slider.
+ */
 export const Tick = Template.bind({});
 Tick.args = {
 	...Default.args,
-	label: undefined,
 	showTicks: true,
 };
 Tick.tags = ["!dev"];
@@ -201,6 +268,23 @@ Tick.parameters = {
 		disableSnapshot: true,
 	},
 };
+
+/**
+ * Spectrum tick slider with tick labels.
+ */
+export const TickWithLabels = Template.bind({});
+TickWithLabels.args = {
+	...Default.args,
+	showTicks: true,
+	showTickLabels: true,
+};
+TickWithLabels.tags = ["!dev"];
+TickWithLabels.parameters = {
+	chromatic: {
+		disableSnapshot: true,
+	},
+};
+TickWithLabels.storyName = "Tick with labels";
 
 export const Disabled = Template.bind({});
 Disabled.args = {
@@ -214,18 +298,21 @@ Disabled.parameters = {
 	},
 };
 
-export const WithFocus = Template.bind({});
-WithFocus.args = {
+export const Focused = Template.bind({});
+Focused.args = {
 	...Default.args,
 	isFocused: true,
 };
-WithFocus.tags = ["!dev"];
-WithFocus.parameters = {
+Focused.tags = ["!dev"];
+Focused.parameters = {
 	chromatic: {
 		disableSnapshot: true,
 	},
 };
 
+/**
+ * A gradient can be added to the track of any slider to give more meaning to the range of values. Tracks with a gradient can also have a fill. A gradient track should not be used for choosing a precise color; use a [color slider](/docs/components-color-slider--docs), [color area](/docs/components-color-area--docs), or [color wheel](/docs/components-color-wheel--docs) instead.
+ */
 export const Gradient = Template.bind({});
 Gradient.args = {
 	...Default.args,
@@ -244,17 +331,21 @@ Gradient.parameters = {
 	},
 };
 
-export const SideLabel = Template.bind({});
-SideLabel.args = {
+/**
+ * Labels can be placed either on top or on the side. Side labels are most useful when vertical space is limited.
+ */
+export const WithSideLabel = Template.bind({});
+WithSideLabel.args = {
 	...Default.args,
 	labelPosition: "side",
 };
-SideLabel.tags = ["!dev"];
-SideLabel.parameters = {
+WithSideLabel.tags = ["!dev"];
+WithSideLabel.parameters = {
 	chromatic: {
 		disableSnapshot: true,
 	},
 };
+WithSideLabel.storyName = "With side label";
 
 // ********* VRT ONLY ********* //
 export const WithForcedColors = SliderGroup.bind({});
@@ -262,6 +353,6 @@ WithForcedColors.tags = ["!autodocs", "!dev"];
 WithForcedColors.parameters = {
 	chromatic: {
 		forcedColors: "active",
-		modes: disableDefaultModes
+		modes: disableDefaultModes,
 	},
 };